+++ src/FunFair.CodeAnalysis/ProhibitedMethodInvocationsDiagnosticsAnalyzer.cs
@@ -151,9 +151,8 @@ namespace FunFair.CodeAnalysis
 
             foreach (IEnumerable<string> ruleSignature in ruleSignatures)
             {
-                methodSignatureList.RemoveAll(match: methodSymbol => methodSymbol
-                                                                     .Parameters.Select(selector: parameterSymbol => SymbolDisplay.ToDisplayString(parameterSymbol.Type))
-                                                                     .SequenceEqual(ruleSignature));
+                methodSignatureList.RemoveAll(match: methodSymbol => methodSymbol.Parameters.Select(selector: parameterSymbol => SymbolDisplay.ToDisplayString(parameterSymbol.Type))
+                                                                                 .SequenceEqual(ruleSignature));
             }
 
             return methodSignatureList;
@@ -176,7 +175,7 @@ namespace FunFair.CodeAnalysis
             {
                 this.SourceClass = sourceClass;
                 this.BannedMethod = bannedMethod;
-                this.Rule = CreateRule(code: ruleId, title: title, message: message);
+                this.Rule = RuleHelpers.CreateRule(code: ruleId, category: CATEGORY, title: title, message: message);
                 this.BannedSignatures = bannedSignatures;
             }
 
@@ -195,20 +194,6 @@ namespace FunFair.CodeAnalysis
             ///     Full qualified name of method
             /// </summary>
             public string QualifiedName => string.Concat(str0: this.SourceClass, str1: ".", str2: this.BannedMethod);
-
-            private static DiagnosticDescriptor CreateRule(string code, string title, string message)
-            {
-                LiteralString translatableTitle = new LiteralString(title);
-                LiteralString translatableMessage = new LiteralString(message);
-
-                return new DiagnosticDescriptor(id: code,
-                                                title: translatableTitle,
-                                                messageFormat: translatableMessage,
-                                                category: CATEGORY,
-                                                defaultSeverity: DiagnosticSeverity.Error,
-                                                isEnabledByDefault: true,
-                                                description: translatableMessage);
-            }
         }
     }
 }